refactor(providers): extract credential lifecycle resolver module (#184)
